requirements.txt
selenium
selenium-3.141.0
page-objects-1.1.0

